moved prepare middleware to its own file
