+++ api/middleware/index.js
@@ -15,10 +15,6 @@ const check = require('../helpers/check')
 const {merge} = require('../helpers/replace')
 const {send_error} = require('../helpers/errors')
 
-const models = require('../models')
-
-const {responses} = require('../../config/resobj')
-
 //SETTINGS
 const unqiue_fields = {
     workouts: ['name'],
@@ -29,62 +25,61 @@ const unqiue_fields = {
     logs: [],
 }
 
+// //:
+// recurssion = async (struct, values) => {
+//     const res = {}
+//     for(const key of Object.keys(struct)) {
+//         const type = Array.isArray(struct[key]) ?  'array' : typeof struct[key]
+//         const tbl = key.split('_')[key.split('_').length-1]
+//         switch(type) {
+//             case 'string': {
+//                 if(struct[key]) res[key] = (await models.get(tbl, {id: values[key]}))[struct[key]]
+//                 else res[key] = values[key]
+//                 break
+//             }
+//             case 'object': {
+//                 const row = await models.get(tbl, {id: values[key]})
+//                 res[key] = await recurssion(struct[key], row)
+//                 break
+//             }
+//             case 'array': {
+//                 const rows = await models.get_all(tbl, {})
+//                 res[key] = []
+//                 switch(typeof struct[key][0]) {
+//                     case 'string': {
+//                         for(const idx in rows) {
+//                             res[key].push((await recurssion({[struct[key][0]]: ''}, rows[idx], tbl))[struct[key][0]])
+//                         }
+//                         break;
+//                     }
+//                     case 'object': {
+//                         for(const idx in rows) {
+//                             const that = await recurssion(struct[key][0], rows[idx], tbl)
+//                             res[key].push(that)
+//                         }
+//                         break;
+//                     }
+//                 }
+//                 break
+//             }
+//         }
+//     }
+//     return res
+// }
 
-//:
-recurssion = async (struct, values) => {
-    const res = {}
-    for(const key of Object.keys(struct)) {
-        const type = Array.isArray(struct[key]) ?  'array' : typeof struct[key]
-        const tbl = key.split('_')[key.split('_').length-1]
-        switch(type) {
-            case 'string': {
-                if(struct[key]) res[key] = (await models.get(tbl, {id: values[key]}))[struct[key]]
-                else res[key] = values[key]
-                break
-            }
-            case 'object': {
-                const row = await models.get(tbl, {id: values[key]})
-                res[key] = await recurssion(struct[key], row)
-                break
-            }
-            case 'array': {
-                const rows = await models.get_all(tbl, {})
-                res[key] = []
-                switch(typeof struct[key][0]) {
-                    case 'string': {
-                        for(const idx in rows) {
-                            res[key].push((await recurssion({[struct[key][0]]: ''}, rows[idx], tbl))[struct[key][0]])
-                        }
-                        break;
-                    }
-                    case 'object': {
-                        for(const idx in rows) {
-                            const that = await recurssion(struct[key][0], rows[idx], tbl)
-                            res[key].push(that)
-                        }
-                        break;
-                    }
-                }
-                break
-            }
-        }
-    }
-    return res
-}
 
+// prepare2 = async (req, res, next) => {
+//     const struct = responses[req.data.table]
+//     const dbres = req.data.response
 
-prepare2 = async (req, res, next) => {
-    const struct = responses[req.data.table]
-    const dbres = req.data.response
-
-    try {
-        req.data.response = await recurssion(struct, dbres)
-    } catch(err) {
-        console.log(err)
-    }
-    console.log('res', req.data.response)
-    next()
-}
+//     try {
+//         req.data.response = await recurssion(struct, dbres)
+//     } catch(err) {
+//         console.log(err)
+//     }
+//     console.log('res', req.data.response)
+//     next()
+// }
 
 //:
 prepare = async (req, res, next) => {
@@ -178,6 +173,5 @@ module.exports = {
     schema,
     id,
     prepare,
-    prepare2,
     encrypt,
 }
